ajouts des changements des données de la page affichageProfil avec les données rentrées par l'utilisateurs

+++ Champolove/src/application/Profil.java
@@ -38,8 +38,8 @@ public class Profil {
 	}
 
 
-	public int compareTo(Profil o) {
-		int res=0;
+	public Double compareTo(Profil o) {
+		Double res=0.0;
 		int totalDesGouts=0;
 		int totalDesHobbies=0;
